add clock in/out and format input and table
Currently the model does not include a clock in/out datetime field
and the index page doesn't show any data. This is a problem because
we need to track the duration and we need to show the data. Fix this
by creating a table to show and styling it with DataTables, and
creating new fields for clocking in/out and using the datetimepicker
with the weird name.