index: fix sorting, add pageLength, align numbers
Currently the sorting of the profitability is incorrect because of the
extra $ and / hour added onto the data. This is a problem because it
won't sort it correctly with this extra text. Fix this by removing it
and also adding a pageLength and aligning the numbers to the right of
the column.