profit: catch divideByZero error
Currently when we are getting the profitability of a Job, it is
possible for the user to enter the same date and time for the clock in
and out fields. This is a problem because it throws an error and ruins
the whole page. Fix this by returning the gross_revenue if the job time
is zero.