fix(profit): sort dates and times correctly
Currently the sorting for dates and times is not working correctly.
This is a problem because it shouldn't be enabled if it doesn't work.
Fix this by using the default sort for date and writing a regex custom
sort for the times since they aren't readable by JS Date parse